refactor(040): migra packages/core/src (index+zodSetup) para TS + corrige package.json exports órfãos
