+++ packages/core/package.json
@@ -4,11 +4,11 @@
   "version": "0.2.0-phase3",
   "type": "module",
   "exports": {
-    ".": "./src/index.js",
-    "./schemas": "./src/schemas/index.js",
-    "./schemas/*": "./src/schemas/*.js",
-    "./utils": "./src/utils/index.js",
-    "./utils/*": "./src/utils/*.js",
-    "./protocols-utils": "./src/protocols-utils/index.js"
+    ".": "./src/index.ts",
+    "./schemas": "./src/schemas/index.ts",
+    "./schemas/*": "./src/schemas/*.ts",
+    "./utils": "./src/utils/index.ts",
+    "./utils/*": "./src/utils/*.ts",
+    "./protocols-utils": "./src/protocols-utils/index.ts"
   }
 }
