+++ packages/core/src/index.ts
@@ -11,7 +11,7 @@
  */
 
 // Aplica locale PT-BR no Zod (efeito colateral ao importar @dosiq/core)
-import './zodSetup.js'
+import './zodSetup'
 
 // Re-exporte de schemas (serao populados em 2.2)
 export * from './schemas/index'
